Hack-tastic browserify-ification.
This is a collection of quick hacks to get Browserify building
our code. There are a lot of subtle ordering requirements due to
global mutation and quasi-namespaces. This resolves only the
conflicts that were blocking the build. The rest are just glued
on to window to call out their evil globalness.